+++ scripts/view-logs.sh
@@ -7,8 +7,6 @@ tail -f logs/*.log | grep --line-buffered "" |
       -e $'s/.*graph-node.log.*/\033[0;33m[SUBGRAPH]\033[0m &/' \
       -e $'s/.*contracts-deploy.log.*/\033[0;32m[CONTRACTS]\033[0m &/' \
       -e $'s/.*subgraph-deploy.log.*/\033[0;34m[SUBGRAPH-DEPLOY]\033[0m &/' \
-      -e $'s/.*arlocal.log.*/\033[1;34m[ARLOCAL]\033[0m &/' \
-      -e $'s/.*offchain-api.log.*/\033[0;32m[OFFCHAIN-API]\033[0m &/' \
       -e $'s/.*temporal-server.log.*/\033[1;35m[TEMPORAL-SERVER]\033[0m &/' \
       -e $'s/.*temporal-processor.log.*/\033[0;35m[TEMPORAL-PROCESSOR]\033[0m &/' \
       -e $'s/.*explorer.log.*/\033[0;33m[EXPLORER]\033[0m &/'